requirements.txt
PyPDF2 
openai==0.28.1 
dotenv
streamlit 
google-generativeai
#selenium
plotly